Remove ruby 1.9.3 support
Ruby 1.9.3 shouldn’t be used anymore since 2.0+ is really stable and
provides a lot of enhancements.

One of them is the `const_get` resolution which makes the tests fail.
Instead of go through the ancestor tree of a constant name I am
removing the 1.9.3 support. If this is really necessary we can fix the 
`const_get` behaviour.